Level 1
Set up Level 1 Tilemaps. Updated the rock throwing script by having the rock travel in the direction of the distance between the rock and the mouse pointer, instead of having it instantiate right at the mouse pointer. Set the mouse pointer visibility in the game to false, and instead the mouse pointer is a rock traced in a yellow-orange color. Fixed the rock circle collider so that collisions with the trees looks better. Added a bounce physics material to the tree tilemap, that way the rock bounces off.